Implementations on the required information
Case study focus on the problem

Challenges of urban mobility in Brazil:
Urban mobility in Brazil faces several types of challenges. However, some of them stand out due to their impact and tendency to worsen.

Overpopulated cities:
According to IBGE estimates, the Brazilian population now exceeds 200 million inhabitants, and many smaller cities have seen an increase well above the average in the last decade. This shows that capitals and metropolitan regions are already too full.
The penetration of services like Uber and 99 has helped with part of the demand. Likewise, the entry of companies offering alternative means of transport, such as bicycles and electric scooters for rent, is another attempt to deal with the situation.

Focus on cars:
Another problem is that a large part of the Brazilian territory was designed for automobiles amid the expansion of highways in the 1950s and 1960s, as well as the growth of the automobile industrial sector.
Brasília is an example of the application of these measures, which ended up restricting the diversity of urban transport. The city was designed only for transportation by car and is bisected by a highway. With the large-scale increase in population, today we face traffic jams and a lack of adequate signage for pedestrians.

Aggression of the ozone layer:
How the use of cars damages the ozone layer
The black list of products harmful to the ozone layer includes nitric and nitrous oxides expelled by vehicle exhausts and CO2 produced by the burning of fossil fuels, such as coal and oil. But, in terms of destructive effects on the ozone layer, nothing compares to the group of gases called chlorofluorocarbons, the CFCs.